move core content to its own folder
closes #321
