+++ projects/ngx-d3timeline/src/lib/axis/axis-utils.ts
@@ -1,7 +1,7 @@
-import { Point } from '../point';
+import { Point } from '../core/point';
 import { Scale } from '../scale-types';
-import { Orientation } from '../orientation';
-import { createLine, Line } from './line';
+import { Orientation } from '../core/orientation';
+import { createLine, Line } from '../core/line';
 import { Axis } from './axis';
 import { TickMarkRenderer } from '../tick-mark/tick-mark-renderer';
 import { getTickMarks } from '../tick-mark/tick-mark-utils';
