+++ projects/ngx-d3timeline/src/lib/content/clip-path.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
-import { Rectangle } from '../rectangle';
+import { Rectangle } from '../core/rectangle';
 
 @Component({
   selector: '[ngx-d3timeline-clip-path]',
