+++ projects/ngx-d3timeline/src/lib/core/orientation.ts
@@ -1,0 +1,10 @@
+export enum Orientation {
+  Horizontal = 'Horizontal',
+  Vertical = 'Vertical'
+}
+
+export function flipOrientation(orientation: Orientation): Orientation {
+  return orientation === Orientation.Vertical
+    ? Orientation.Horizontal
+    : Orientation.Vertical;
+}
