+++ projects/ngx-d3timeline/src/lib/core/rectangle.ts
