+++ projects/ngx-d3timeline/src/lib/core/text-utils.ts
