+++ projects/ngx-d3timeline/src/lib/core/types.ts
