+++ projects/ngx-d3timeline/src/lib/options-utils.ts
@@ -1,0 +1,6 @@
+import { Orientation } from './core/orientation';
+import { Options } from './options';
+
+export function getTimeOrientation(options: Options): Orientation {
+  return options && Orientation[options.orientation];
+}
