+++ projects/ngx-d3timeline/src/lib/options.selectors.ts
@@ -1,6 +1,7 @@
 import { createSelector } from './store-lib/selector/create-selector';
 import { selectOptions } from './store/state';
-import { flipOrientation, getTimeOrientation } from './orientation-utils';
+import { flipOrientation } from './core/orientation';
+import { getTimeOrientation } from './options-utils';
 
 export const selectTimeOrientation = createSelector(
   selectOptions,
