--- projects/ngx-d3timeline/src/lib/orientation.ts
@@ -1,4 +1,0 @@
-export enum Orientation {
-  Horizontal = 'Horizontal',
-  Vertical = 'Vertical'
-}
