+++ projects/ngx-d3timeline/src/lib/store/reducer.ts
@@ -2,7 +2,7 @@ import { State, initialState } from './state';
 import { Actions, ActionType } from './actions';
 import { View } from '../view/view';
 import { TimelineDragEvent } from '../content/timeline-drag-event';
-import { identifier } from '../types';
+import { identifier } from '../core/types';
 import { initialisePositionedActivity } from '../positioned-activity';
 import { defaultOptions, CompleteOptions } from '../options';
