+++ projects/ngx-d3timeline/src/lib/store/state.ts
@@ -2,7 +2,7 @@ import { View } from '../view/view';
 import { TimelineDragEvent } from '../content/timeline-drag-event';
 import { createSliceSelector } from '../store-lib/selector/slice-selector';
 import { PositionedActivity } from '../positioned-activity';
-import { identifier } from '../types';
+import { identifier } from '../core/types';
 import { defaultOptions, CompleteOptions } from '../options';
 
 export interface State {
