+++ projects/ngx-d3timeline/src/lib/tick-mark/resource-axis-tick-mark-renderer.ts
@@ -3,7 +3,7 @@ import {
   TickMarkRenderer,
   tickLabelSpacing
 } from '../tick-mark/tick-mark-renderer';
-import { Orientation } from '../orientation';
+import { Orientation } from '../core/orientation';
 
 export function getResourceAxisTickMarkRenderer(
   scale: BandScale,
