--- projects/ngx-d3timeline/src/lib/transform-utils.ts
@@ -1,5 +1,0 @@
-import { Point } from './point';
-
-export function pointToTransform(point: Point): string {
-  return `translate(${point.x}, ${point.y})`;
-}
